Create ryc5nfd.Rmd
The purpose of this code is to demonstrate the reinforcement of my learning with KMeans clustering and ggplot visualization.

If the code is run, multiple figures that I have created can be seen.

One thing I was curious about that I looked up after class was the evaluation of the best number of k clusters. Specifically, I looked up what exactly were the Hubert Statistic Values. In addition, I googled what an area plot was, because I used it in my example without actually knowing what it was.