Add files via upload
The purpose of this code was to pracitce making a KMeans Clustering model of the data provided.  I used 5 clusters as determined by the Elbow Plot, and I colored based on the variable "Won" to see if there were any patterns between Winning and Correlation A and C.  Overall, it appeared that the Won variable, TRUE or FALSE, was spread evenly throughout the 5 clusters.

One thing I looked up after class was why the KMeans algorithm would converge when setting the seed (starting the algorithm) at one point, when it wouldn't at another.  I learned that this is because convergence depends on the initial centroids chosen for the KMeans.  Some initial centroid positions will converge more quickly than others, while others will not converge at all.  Therefore, for the seed I chose, 1984, the centroids did converge.  However, other seeds will not always converge depending on the initial centroids.