Week11Lab Submission for Claire Dozier (ced9mq)
This is my homework submission for the Week 11 Reinforcement Lab. In this lab, I chose to focus on clustering and evaluation. I used the numeric variables in the dataset to create clusters. I evaluated the variance explained using five clusters, but also created an elbow plot to validate my cluster choice. I also used ggplot visualizations to understand different clusters and how they relate to other variables in the dataset.