Create njp2cyf.R
The purpose of this code was to reinforce topics that we have so far this semester in this class. Specifically, I chose to reinforce topics such as Kmeans, Clustering, and ggplot, because I feel like I could improve my skills in those areas. Therefore, I ran Kmeans on the data set provided, performed clustering, and created ggplots that demonstrate the clustering of the data. 

I could not figure out how you wanted us to upload a figure generated, but the code above produces a few interesting figures. 

After class, I decided to look up information about KNN, because I struggled to complete the KNN lab, and I now feel more comfortable with the topic. I look forward to applying the information that I have learned to the KNN lab, so that I can improve the results. The link that I looked at is the following: https://towardsdatascience.com/machine-learning-basics-with-the-k-nearest-neighbors-algorithm-6a6e71d01761