Add files via upload
This code contains practice of clustering using the correlation data that we were provided in class. I have also attached the graph that was produced, where we cluster based on correlation and color based on whether the "won" variable was true or false. We can kind of see that there does not seem to be much relation between each of the 5 clusters and whether that observation was True or False for the one variable, but it was still good practice to make the visualization and practice kmeans and all of that. Additionally, kind of a funny thing that my group worked through was relating to 3D visualization. We got really excited to try and create a 3D graph of our clusters, but as soon as we did this, we quickly realized that we couldn't make a 3D visualization because we were only clustering based on 2 variables and not 3!
One thing I looked up after class was about when clustering is not a good thing to use. While we were conducting our analysis, I wanted to find a way to add in some extra variable so that we could make a 3D graph. However, I read about how clustering is not effective when you are dealing with a lot of categorical variables or binary variables (as is the case with this dataset). That being said, I still think I have enough information that this data is good for clustering. Even though we only have 2 continuous variables to deal with, by making an initial graph it is very apparent that groups appear. And, based on the reading that I did, this can often be grounds enough for clustering to apply.